- Move Log4NetLogger.cs from GoogleApis.DotNet4 to GoogleApis.Tests. - Remove log4net dependency from GoogleApis.DotNet4
