+++ Src/Support/GoogleApis.DotNet4/GoogleApis.PlatformServices_Net45.csproj
@@ -66,7 +66,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Apis\Logging\Log4NetLogger.cs" />
     <Compile Include="Apis\Util\Store\FileDataStore.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="..\CommonAssemblyInfo.cs">
