+++ Src/Support/GoogleApis.Tests/GoogleApis.Tests.csproj
@@ -117,6 +117,7 @@
     <Compile Include="Apis\Utils\RepeatableTest.cs" />
     <Compile Include="ApplicationContextTests.cs" />
     <Compile Include="Json\JsonExplicitNullTest.cs" />
+    <Compile Include="Logging\Log4NetLogger.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="..\CommonAssemblyInfo.cs">
